requirements.txt
graphql-core==3.2.3
pydantic==1.10.4
strawberry-graphql==0.154.0
mypy==0.991